requirements.txt
biopython=1.83
requests=2.31.0